+++ admin/includes/class-user.php
@@ -33,7 +33,7 @@ class User {
 		global $rs_query;
 		
 		// Set up pagination
-		$page = paginate((int)($_GET['page'] ?? 1));
+		$page = paginate((int)($_GET['paged'] ?? 1));
 		?>
 		<div class="heading-wrap">
 			<h1>Users</h1>